Parse PRG ROM data into an array of NSData

+++ MOSSimTests/NES/Test Files/NESTestFile.h
@@ -1,2 +1,2 @@
 extern const unsigned char NESTestFile[];
-extern const int NESTestFileLength;
+extern const NSUInteger NESTestFileLength;
